+++ server/morak_back_end/src/test/java/com/morakmorak/morak_back_end/controller/auth_controller/AuthControllerTest.java
@@ -1,0 +1,51 @@
+package com.morakmorak.morak_back_end.controller.auth_controller;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.morakmorak.morak_back_end.controller.AuthController;
+import com.morakmorak.morak_back_end.controller.ExceptionController;
+import com.morakmorak.morak_back_end.mapper.UserMapper;
+import com.morakmorak.morak_back_end.security.resolver.JwtArgumentResolver;
+import com.morakmorak.morak_back_end.security.util.JwtTokenUtil;
+import com.morakmorak.morak_back_end.service.AuthService;
+import com.morakmorak.morak_back_end.util.SecurityTestConfig;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Import;
+import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static com.morakmorak.morak_back_end.util.SecurityTestConstants.ACCESS_TOKEN;
+import static com.morakmorak.morak_back_end.util.SecurityTestConstants.REFRESH_TOKEN;
+
+@WithMockUser
+@Import(SecurityTestConfig.class)
+@WebMvcTest({AuthController.class, UserMapper.class, ExceptionController.class})
+@MockBean(JpaMetamodelMappingContext.class)
+@AutoConfigureRestDocs(uriScheme = "https", uriHost = "docs.api.com")
+public class AuthControllerTest {
+    @Autowired
+    MockMvc mockMvc;
+
+    @Autowired
+    UserMapper userMapper;
+
+    @Autowired
+    ObjectMapper objectMapper;
+
+    @MockBean
+    AuthService authService;
+
+    @MockBean
+    JwtArgumentResolver jwtArgumentResolver;
+
+    JwtTokenUtil jwtTokenUtil;
+
+    @BeforeEach
+    public void init() {
+        jwtTokenUtil = new JwtTokenUtil(ACCESS_TOKEN, REFRESH_TOKEN);
+    }
+}
