+++ server/morak_back_end/src/test/java/com/morakmorak/morak_back_end/dto/auth_dto/EmailDtoTest.java
@@ -1,5 +1,6 @@
-package com.morakmorak.morak_back_end.dto;
+package com.morakmorak.morak_back_end.dto.auth_dto;
 
+import com.morakmorak.morak_back_end.dto.EmailDto;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
